áda

+++ APP0200025/Controllers/ChoTiepNhanHoSoController.cs
@@ -410,7 +410,7 @@ namespace APP0200025.Controllers
             return RedirectToAction("Index", models);
         }
 
-        public PartialViewResult HangHoaChiTiet(int iID_MaHangHoa)
+        public PartialViewResult HangHoaChiTiet(long iID_MaHangHoa)
         {
             var hangHoa = clHangHoa.GetHangHoaById(iID_MaHangHoa);
             return PartialView("~/Views/Partial/Partial_ThongTinChiTietTACN.cshtml", hangHoa);
