+++ DATA0200025/Models/GiayChungNhanHopQuyModels.cs
@@ -9,8 +9,8 @@ namespace DATA0200025.Models
   public  class GiayChungNhanHopQuyModels
     {
         public int iID_ChungNhanHopQuy { get; set; }
-        public int iID_MaHoSo { get; set; }
-        public int iID_MaHangHoa { get; set; }
+        public long iID_MaHoSo { get; set; }
+        public long iID_MaHangHoa { get; set; }
         public string sSoChungNhan { get; set; }
         public DateTime dNgayCap { get; set; }
         public string sTenFile { get; set; }
